# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: PubliMentor API
  version: 1.0.0
  description: |
    Scientific Editorial Workflow Management API.
    
    ## Overview
    PubliMentor provides APIs for managing manuscripts, finding reviewers,
    checking conflicts of interest, and validating research integrity.
    
    ## Authentication
    All endpoints except health checks require Bearer token authentication.
    
    ```
    Authorization: Bearer <access_token>
    ```
    
    ## Rate Limiting
    API requests are rate limited. When exceeded, responses include:
    - `429 Too Many Requests` status
    - `Retry-After` header with seconds until reset
    
    ## Request IDs
    All responses include `X-Request-ID` header for tracing.
    Include `X-Request-ID` in requests to correlate logs.
  contact:
    name: PubliMentor Support
    email: support@publimentor.io
  license:
    name: Proprietary
    url: https://publimentor.io/license
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://api.publimentor.io/api
    description: Production server
...
paths:
...
  /manuscripts:
    get:
      tags: [Manuscripts]
      summary: List manuscripts
      description: Get paginated list of manuscripts accessible to the user
      operationId: listManuscripts
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/ManuscriptStatus'
        - name: journalId
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of manuscripts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManuscriptListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /manuscripts/upload:
    post:
      tags: [Manuscripts]
      summary: Upload a manuscript
      description: Upload a PDF, DOCX, or LaTeX document
      operationId: uploadManuscript
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, publisherId]
              properties:
                file:
                  type: string
                  format: binary
                  description: The manuscript file (PDF, DOCX, or LaTeX)
                publisherId:
                  type: string
                  description: Publisher ID for the manuscript
                journalId:
                  type: string
                  description: Optional journal ID
      responses:
        '201':
          description: Manuscript uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManuscriptResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  responses:
    BadRequest:
      description: Bad request - validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    RateLimited:
      description: Rate limit exceeded
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds until rate limit resets
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    ApiResponse:
      type: object
      required: [success, requestId, timestamp]
      properties:
        success:
          type: boolean
        requestId:
          type: string
        timestamp:
          type: string
          format: date-time
        data:
          type: object
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    ErrorResponse:
      type: object
      required: [success, requestId, timestamp, error]
      properties:
        success:
          type: boolean
          enum: [false]
        requestId:
          type: string
        timestamp:
          type: string
          format: date-time
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            target:
              type: string
            details:
              type: array
              items:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
                  target:
                    type: string
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        pageSize:
          type: integer
        totalCount:
          type: integer
        totalPages:
          type: integer
        hasMore:
          type: boolean

    # Health schemas
...
    ManuscriptStatus:
      type: string
      enum: [UPLOADED, EXTRACTING, EXTRACTED, PROCESSING, EMBEDDING, READY, ERROR]
    ManuscriptResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/Manuscript'
    ManuscriptListResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/ManuscriptSummary'
    Manuscript:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        abstract:
          type: string
        status:
          $ref: '#/components/schemas/ManuscriptStatus'
        fileName:
          type: string
        fileType:
          type: string
        wordCount:
          type: integer
        pageCount:
          type: integer
        keywords:
          type: array
          items:
            type: string
        authors:
          type: array
          items:
            $ref: '#/components/schemas/Author'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ManuscriptSummary:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        status:
          $ref: '#/components/schemas/ManuscriptStatus'
        fileName:
          type: string
        createdAt:
          type: string
          format: date-time
    Author:
      type: object
      properties:
        fullName:
          type: string
        email:
          type: string
        orcid:
          type: string
        isCorresponding:
          type: boolean
        affiliations:
          type: array
          items:
            type: string

    # Reviewer schemas
...
security:
  - bearerAuth: []